#2 - updating de ProductsService to check if a product is being used and updating the tests

+++ products/products/service.py
@@ -4,7 +4,7 @@ from nameko.events import event_handler
 from nameko.rpc import rpc
 
 from products import dependencies, schemas
-
+from products.exceptions import ProductInUse
 
 logger = logging.getLogger(__name__)
 
@@ -14,6 +14,8 @@ class ProductsService:
     name = 'products'
 
     storage = dependencies.Storage()
+
+    orders_rpc = RpcProxy('orders')
 
     @rpc
     def get(self, product_id):
@@ -32,6 +34,8 @@ class ProductsService:
     
     @rpc
     def delete(self, product_id):
+        if self._is_product_in_use(product_id):
+            raise ProductInUse(f'Product with id {product_id} is in use and cannot be deleted')
         self.storage.delete(product_id)
 
     @event_handler('orders', 'order_created')
@@ -39,3 +43,6 @@ class ProductsService:
         for product in payload['order']['order_details']:
             self.storage.decrement_stock(
                 product['product_id'], product['quantity'])
+
+    def _is_product_in_use(self, product_id):
+        return len(self.orders_rpc.list_orders_by_product_id(product_id)) > 0
